+++ uvdat/core/rest/network.py
@@ -1,7 +1,7 @@
 from rest_framework.viewsets import ModelViewSet
 
 from uvdat.core.models import Network, NetworkEdge, NetworkNode
-from uvdat.core.rest.guardian import GuardianFilter, GuardianPermission
+from uvdat.core.rest.access_control import GuardianFilter, GuardianPermission
 from uvdat.core.rest.serializers import (
     NetworkEdgeSerializer,
     NetworkNodeSerializer,
